backend/requirements.txt
fastapi
uvicorn
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openpyxl
python-dotenv
